FIX: Back up the finished indices only
If we back up the indices of loop which is executing, it may write
the same data in the same memory space duplicately with different
version number.

Thus, we change the design and back up the finished indices of loop.

TODO: back up the node's input data and output data to NVM.